+++ hiperion_web/src/main/java/ec/com/avila/emision/web/backings/LucroCesanteIncendioBacking.java
@@ -160,7 +160,7 @@ public class LucroCesanteIncendioBacking implements Serializable {
 	 */
 	public List<SelectItem> getFormaSeguroItems() throws HiperionException {
 		this.formaSeguroItems = new ArrayList<SelectItem>();
-		Catalogo catalogo = catalogoService.consultarCatalogoById(HiperionMensajes.getInstancia().getInteger(
+		Catalogo catalogo = catalogoService.consultarCatalogoById(HiperionMensajes.getInstancia().getLong(
 				"ec.gob.avila.hiperion.recursos.catalogoFormaSeguro"));
 		List<DetalleCatalogo> formasSeguro = catalogo.getDetalleCatalogos();
