+++ hiperion_web/src/main/java/ec/com/avila/emision/web/backings/SoatBacking.java
@@ -92,7 +92,6 @@ public class SoatBacking implements Serializable {
 		}
 
 	}
-	
 
 	/**
 	 * 
@@ -133,7 +132,7 @@ public class SoatBacking implements Serializable {
 	public List<SelectItem> getTipoVehiculoItems() throws HiperionException {
 
 		this.tipoVehiculoItems = new ArrayList<SelectItem>();
-		Catalogo catalogo = catalogoService.consultarCatalogoById(HiperionMensajes.getInstancia().getInteger(
+		Catalogo catalogo = catalogoService.consultarCatalogoById(HiperionMensajes.getInstancia().getLong(
 				"ec.gob.avila.hiperion.recursos.catalogoTipoVehiculoSoat"));
 		List<DetalleCatalogo> tipos = catalogo.getDetalleCatalogos();
 
@@ -264,7 +263,7 @@ public class SoatBacking implements Serializable {
 	public List<SelectItem> getTipoSoatItems() throws HiperionException {
 
 		this.tipoSoatItems = new ArrayList<SelectItem>();
-		Catalogo catalogo = catalogoService.consultarCatalogoById(HiperionMensajes.getInstancia().getInteger(
+		Catalogo catalogo = catalogoService.consultarCatalogoById(HiperionMensajes.getInstancia().getLong(
 				"ec.gob.avila.hiperion.recursos.catalogoTipoSoat"));
 		List<DetalleCatalogo> tipos = catalogo.getDetalleCatalogos();
